+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 
 ## Release 0.0.1
 
++   Implemented DEX Contracts
 +   add Changelog